% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_intervals.R
\name{ci_beta}
\alias{ci_beta}
\title{Creates a confidence interval for the standardized regression coefficient \eqn{\beta}.}
\usage{
ci_beta(t, df1, df2, r2, r2x, conf = 0.95, fixed = FALSE, iter = 50)
}
\arguments{
\item{t}{The t-test for the standardized regression coefficient.}

\item{df1}{The numerator degrees of freedom for the regression model.}

\item{df2}{The denominator degrees of freedom for the regression model.}

\item{r2}{The squared multiple correlation for the full regression model.}

\item{r2x}{The coefficient of determination for the predictor of interest. This is the squared multiple correlation when the predictor of interest is predicted by all of the other variables in the regression model.  1 - tolerance.}

\item{conf}{The desired (1-\eqn{\alpha}) confidence level. The default is set to .95.}

\item{fixed}{Specifies whether the predictors in the regression model are either fixed or random. The default is random predictors.}

\item{iter}{Number of iterations for the stochastic approximation optimization. Default is 50.}
}
\value{
Returns the lower and upper endpoints of the confidence interval.
}
\description{
Creates a confidence interval for the standardized regression coefficient \eqn{\beta}.
}
