% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_intervals.R
\name{ci_d}
\alias{ci_d}
\title{Creates a confidence interval for the standardized mean difference. Assumes predictors are fixed.}
\usage{
ci_d(d, n1, n2, conf = 0.95, iter = 50)
}
\arguments{
\item{d}{The estimated standardized mean difference.}

\item{n1}{The sample size for group 1.}

\item{n2}{The sample size for group 2.}

\item{conf}{The desired (1-\eqn{\alpha}) confidence level. Default is .95.}

\item{iter}{Number of iterations for the stochastic approximation optimization. Default is 50.}
}
\value{
Returns the lower and upper endpoints of the confidence interval.
}
\description{
Creates a confidence interval for the standardized mean difference. Assumes predictors are fixed.
}
