% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_intervals.R
\name{ci_f}
\alias{ci_f}
\title{Creates a confidence interval for Cohen's f.}
\usage{
ci_f(f, df1, df2, fixed = FALSE, conf = 0.95)
}
\arguments{
\item{f}{Estimated value of Cohen's f.}

\item{df1}{The degrees of freedom numerator for the F-test associated with f.}

\item{df2}{The degrees of freedom denominator for the F-test associated with f.}

\item{fixed}{Specifies whether the predictors in the model are either fixed or random. The default is FALSE for random predictors.}

\item{conf}{The desired (1-\eqn{\alpha}) confidence level. Default is .95.}
}
\value{
Returns the lower and upper endpoints of the confidence interval.
}
\description{
Creates a confidence interval for Cohen's f.
}

