% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_intervals.R
\name{ci_r}
\alias{ci_r}
\title{Creates a confidence interval for the correlation.}
\usage{
ci_r(r, df, fixed = FALSE, conf = 0.95, iter = 50)
}
\arguments{
\item{r}{The estimated correlation or partial correlation.}

\item{df}{The degrees of freedom.}

\item{fixed}{Specifies whether the variable in the regression model is fixed or random. The default is for random predictor.}

\item{conf}{The desired (1-\eqn{\alpha}) confidence level. Default is .95.}

\item{iter}{Number of iterations for the stochastic approximation optimization. Default is 50.}
}
\value{
Returns the lower and upper endpoints of the confidence interval.
}
\description{
Creates a confidence interval for the correlation.
}

