% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_intervals.R
\name{ci_r2}
\alias{ci_r2}
\title{Creates a confidence interval for the squared multiple correlation.}
\usage{
ci_r2(r2, df1, df2, conf = 0.95, fixed = FALSE)
}
\arguments{
\item{r2}{The estimated squared multiple correlation.}

\item{df1}{The numerator degrees of freedom.}

\item{df2}{The denominator degrees of freedom.}

\item{conf}{The desired (1-\eqn{\alpha}) confidence level. Default is .95.}

\item{fixed}{Specifies whether the predictors in the regression model are either fixed or random. The default is random predictors.}
}
\value{
Returns the lower and upper endpoints of the confidence interval.
}
\description{
Creates a confidence interval for the squared multiple correlation.
}
