% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desired_ep_Ftest.R
\name{desired_ep_Ftest}
\alias{desired_ep_Ftest}
\title{Determines the required sample size in a future study to achieve the desired expected power (ep)
based on the uncertainty associated with an existing study.
Uses the F-test of the previous study as input in the sample size planning process.}
\usage{
desired_ep_Ftest(
  Ftest,
  df1,
  df2,
  desired_ep = 0.8,
  alpha = 0.05,
  filter = 0,
  upper_null = 0,
  estimate_fixed = TRUE,
  future_fixed = TRUE
)
}
\arguments{
\item{Ftest}{The F-test of the previous study.}

\item{df1}{The numerator degrees of freedom.}

\item{df2}{The denominator degrees of freedom.}

\item{desired_ep}{The desired expected power for the future study.}

\item{alpha}{The significance level. Default is \eqn{\alpha = .05}.}

\item{filter}{The filter value reflects the probability of nonsignificant results being filtered.
filter = 0 means that there is no filtering and you would have observed nonsignificant results.
filter = 1 means that only significant results are observed and you would never have seen nonsigificant results if they had occurred.
Filtering is based on alpha = .05 and assumes that are have observed a significant result.
Filtering is conducted by weighting (actually filtering) the posterior distribution.
For instance, if filter = 1, then the posterior of the null (i.e., the noncentrality parameter is 0)
is up to 20 times more likely than when the noncentrality parameter is very large. Setting filter > 0 slows estimation.}

\item{upper_null}{Specifies the upper value of the composite null hypothesis in units of Cohen's f.
The default value of upper_null = 0 keeps the point null hypothesis.
A value of, for instance, upper_null = .05 would remove all posterior values between -.05 and .05 from consideration when calculating expected power.}

\item{estimate_fixed}{Specifies whether the predictor in the regression model is either fixed or random. The default is FALSE for random predictors.}

\item{future_fixed}{Specifies whether the future study will have fixed predictors.}
}
\value{
Returns (1) the sample size required for the future study to achieve the specified level of expected power.
This reflects the uncertainty associated with the previous study and (2) the median 95% confidence interval width for the correlation in the prospective study.
}
\description{
Determines the required sample size in a future study to achieve the desired expected power (ep)
based on the uncertainty associated with an existing study.
Uses the F-test of the previous study as input in the sample size planning process.
}
\examples{
\dontrun{
desired_ep_Ftest(Ftest = 5.0, df1=2, df2=50, desired_ep = 0.80)
}
}
