% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data_lm_functions.R
\name{em.summary}
\alias{em.summary}
\title{Function to return the EM means and standard deviations using the norm package.}
\usage{
em.summary(data)
}
\arguments{
\item{data}{The dataset for the analysis.}
}
\value{
Returns (1) the EM means and (2) the EM standard deviations for the dataset.
}
\description{
Function to return the EM means and standard deviations using the norm package.
}
\examples{
\dontrun{
x <- c(1,2,3,4,5,NA,NA,7,7,7,7)
y <- c(2.1,NA,2.1,1.8,2,2.2,4,NA,7,7,7)
temp_data <- as.data.frame(cbind(y, x))
em.summary(temp_data)
}
}
