% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_grpFabs.r
\name{GFabs}
\alias{GFabs}
\title{A Group Forward and Backward Stagewise (GFabs) algorithm for Group penalized problem.}
\usage{
GFabs(
  W,
  y,
  group,
  status = NULL,
  sigma = NULL,
  weight = NULL,
  model = c("spr", "square", "cox", "logistic"),
  back = TRUE,
  stoping = TRUE,
  eps = 0.01,
  xi = 10^-6,
  iter = 10^4,
  lambda.min = NULL
)
}
\arguments{
\item{W}{The design matrix.}

\item{y}{The survival outcome.}

\item{group}{The grouping vector.}

\item{status}{The censoring indicator.}

\item{sigma}{The smoothing parameter in SPR.}

\item{weight}{The weight vector of groups.}

\item{model}{The loss function used.}

\item{back}{The indicator of whether to take backward steps.}

\item{stoping}{The indicator of whether to stop iteration when lambda is less than lambda.min.}

\item{eps}{The step size for GFabs.}

\item{xi}{The threshhold for GFabs.}

\item{iter}{The maximum number of outer-loop iterations allowed.}

\item{lambda.min}{The smallest value for lambda, as a fraction of lambda.max.}
}
\value{
A list.
\itemize{
  \item Beta - The standardized estimation of covariates.
  \item beta - The optimal standardized estimation of covariates.
  \item lambda - Lambda sequence.
  \item direction - Direction of GFabs.
  \item active - Active set for each step.
  \item iter - Iterations.
  \item BIC - The bic for each solution.
  \item group - The grouping vector.
  \item opt - Position of the optimal tuning based on BIC.
}
}
\description{
A Group Forward and Backward Stagewise (GFabs) algorithm for Group penalized problem.
}
\examples{
library(mvtnorm)
sigma = outer(1:20, 1:20, FUN = function(x, y) 0.3^(abs(x - y)))
x     = rmvnorm(100, mean = rep(0,20), sigma = sigma)
b     = c(5, -5, 5, -5, rep(0, 16))
error = c(0.7*rnorm(100)+0.3*rcauchy(100))
y     = x \%*\% b + error
group = seq(1, 20)
fit   <- GFabs(x, y, group)
}
