% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_distributions.R
\name{posterior_r}
\alias{posterior_r}
\title{Samples from the posterior distribution of the correlation coefficient. The posterior distribution is
generated using Fisher's fiducial approach and corresponds exactly to the use of a Jeffrey's prior.}
\usage{
posterior_r(r, df, filter = 0, upper_null = 0, fixed = FALSE,
  ndraws = 2e+05)
}
\arguments{
\item{r}{The correlation of the previous study.}

\item{df}{The degrees of freedom in the previous study.}

\item{filter}{The filter value reflects the probability of nonsignificant results being filtered.
filter = 0 means that there is no filtering and you would have observed nonsignificant results.
filter = 1 means that only significant results are observed and you would never have seen nonsigificant results if they had occurred.
Filtering is based on alpha = .05 and assumes that are have observed a significant result.
Filtering is conducted by weighting (actually filtering) the posterior distribution.
For instance, if filter = 1, then the posterior of the null (i.e., the noncentrality parameter is 0)
is up to 20 times more likely than when the noncentrality parameter is very large. Setting filter > 0 slows estimation.}

\item{upper_null}{Specifies the upper value of the composite null hypothesis in units of the correlation.
The default value of upper_null = 0 keeps the point null hypothesis.
A value of, for instance, upper_null = .05 would remove all posterior values between -.05 and .05.}

\item{fixed}{Specifies whether the predictors in the model are either fixed or random. The default is FALSE for random predictors.}

\item{ndraws}{Specifies the number of initial samples from the posterior. For small effect sizes and when filter or upper_null > 0
the number of returned samples from the posterior distribution will be lower than ndraws.}
}
\value{
A vector of samples from the posterior distribution in units of the correlation parameter \eqn{\rho}.
}
\description{
Samples from the posterior distribution of the correlation coefficient. The posterior distribution is
generated using Fisher's fiducial approach and corresponds exactly to the use of a Jeffrey's prior.
}
