% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_distributions.R
\name{posterior_t}
\alias{posterior_t}
\title{Samples from the posterior distribution of the noncentral t-distribution.
The posterior distribution is generated using Fisher's fiducial approach.}
\usage{
posterior_t(t, df, fixed = FALSE, ndraws = 2e+05)
}
\arguments{
\item{t}{The t-test of the previous study.}

\item{df}{The degrees of freedom associated with the t-test.}

\item{fixed}{Is the predictor fixed or random? Default is FALSE to indicate randomly sampled predictor.}

\item{ndraws}{Specifies the number of initial samples from the posterior. For small effect sizes and when filter or upper_null > 0
the number of returned samples from the posterior distribution will be lower than ndraws.}
}
\value{
A vector of samples from the posterior distribution of the noncentral t-distribution.
}
\description{
Samples from the posterior distribution of the noncentral t-distribution.
The posterior distribution is generated using Fisher's fiducial approach.
}
