% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data_lm_functions.R
\name{scale.all.variables}
\alias{scale.all.variables}
\title{Function to standardize all variables based on supplied mean and sd vectors}
\usage{
\method{scale}{all.variables}(data, mean, sd)
}
\arguments{
\item{data}{The dataset}

\item{mean}{Vector of means.}

\item{sd}{Vector of standard deviations}
}
\value{
Returns a dataset where all of the variables have been scaled.
}
\description{
Function to standardize all variables based on supplied mean and sd vectors
}
\examples{
\dontrun{
x <- c(1,2,3,4,5)
y <- c(2.1,2.5,4,5,6)
temp_data <- as.data.frame(cbind(y, x))
scale.all.variables(data=temp_data, mean=sapply(temp_data, mean), sd=sapply(temp_data, sd))
}
}
