% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard.r
\name{standard}
\alias{standard}
\title{Within group standardization}
\usage{
standard(W, group)
}
\arguments{
\item{W}{A matrix.}

\item{group}{The grouping vector.}
}
\value{
A list
\itemize{
  \item xx - The standardized matrix.
  \item center - The colmeans of the original matrix.
  \item scale - A list whose i^{th} element is the transformation matrix of the i^{th} group.
}
}
\description{
Within group standardization
}
\examples{
W = matrix(rnorm(200), 10, 20)
group <- rep(c(1:10), each = 2)
W_tilde <- standard(W, group)
}
